packages/mop_utils/requirements.txt
azureml-contrib-services==1.48
pyraisdk==0.1.0
PyYAML>=6.0
numpy